We added elements of modern c++ and worked on the appearance of the code.

+++ Phase 10/Phase.h
@@ -1,5 +1,7 @@
 #pragma once
+
 #include <vector>
+
 #include "Player.h"
 
 class Card;
@@ -17,16 +19,16 @@ private:
    // void ModifyHandCards2(Player& player, std::vector<int> options1, std::vector<int> options2);
 
 
-	bool IsPhase1(std::vector<Card> cards1, std::vector<Card> cards2, int& incorrect );
-	bool IsPhase2(std::vector<Card> cards1, std::vector<Card> cards2, int& incorrect);
-	bool IsPhase3(std::vector<Card> cards1, std::vector<Card> cards2, int& incorrect); 
+	bool IsPhase1(std::vector<Card> cards1, std::vector<Card> cards2, uint8_t& incorrect );
+	bool IsPhase2(std::vector<Card> cards1, std::vector<Card> cards2, uint8_t& incorrect);
+	bool IsPhase3(std::vector<Card> cards1, std::vector<Card> cards2, uint8_t& incorrect);
 	bool IsPhase4(std::vector<Card> cards); 
 	bool IsPhase5(std::vector<Card> cards);
 	bool IsPhase6(std::vector<Card> cards);
-	bool IsPhase7(std::vector<Card> cards1, std::vector<Card> cards2, int& incorrect);
+	bool IsPhase7(std::vector<Card> cards1, std::vector<Card> cards2, uint8_t& incorrect);
 	bool IsPhase8(std::vector<Card> cards);
-	bool IsPhase9(std::vector<Card> cards1, std::vector<Card> cards2, int& incorrect); 
-	bool IsPhase10(std::vector<Card> cards1, std::vector<Card> cards2, int& incorrect); 
+	bool IsPhase9(std::vector<Card> cards1, std::vector<Card> cards2, uint8_t& incorrect);
+	bool IsPhase10(std::vector<Card> cards1, std::vector<Card> cards2, uint8_t& incorrect);
 
 };
